RTI-958 Add healtcheck timeout to proxy reload timeout when healthchecks are ignored

+++ helper/types.go
@@ -24,7 +24,7 @@ import (
 )
 
 type DeployerConfig struct {
-	HealthTimeout       int64
+	HealthTimeout       int
 	K8sClient           *k8s.K8sClient
 	EtcdRegistry        *etcdregistry.EtcdRegistry
 	ProxyConfigurator   *proxies.ProxyConfigurator
